:art: Refractor `storage_helpers` module
The `storage_helpers` module is now appropiately divided into various
modules according to what function it serves. Including `whitelist` and
`treasury_cacher` logic.
