+++ backend/app/libs/tasks/get_assets.py
@@ -12,13 +12,18 @@ from httpx import ReadTimeout
 
 from ... import db
 from ...celery_main import app as celery_app
-from ...treasury import build_treasury_with_assets
-from .. import price_stats
-from ..storage_helpers import (
-    get_and_store_treasury_list,
-    retrieve_treasuries_metadata,
+from ...token_whitelists import (
     store_and_get_covalent_pairs_whitelist,
     store_and_get_tokenlist_whitelist,
+)
+from ...treasury import (
+    build_treasury_with_assets,
+    get_treasury_list,
+    retrieve_treasuries_metadata,
+    store_treasuries_metadata,
+)
+from .. import price_stats
+from .redis import (
     store_asset_correlations,
     store_asset_hist_balance,
     store_asset_hist_performance,
@@ -62,7 +67,7 @@ def reload_treasuries_stats():
 
     treasuries = retrieve_treasuries_metadata(db)
     if not treasuries:
-        get_and_store_treasury_list(db)
+        store_treasuries_metadata(db, get_treasury_list())
         treasuries = retrieve_treasuries_metadata(db)
 
     for treasury_metadata in treasuries:
@@ -145,4 +150,4 @@ def reload_whitelist():
 
 @celery_app.task(name="tasks.reload_treasuries_list")
 def reload_treasuries_list():
-    get_and_store_treasury_list(db)
+    store_treasuries_metadata(db, get_treasury_list())
